myapplication -> pyculator
