+++ app/src/main/AndroidManifest.xml
@@ -10,13 +10,13 @@
         android:label="@string/app_name"
         android:name="com.chaquo.python.android.PyApplication"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MyApplication"
+        android:theme="@style/Theme.pyculator"
 
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.MyApplication"
+            android:theme="@style/Theme.pyculator"
             android:windowSoftInputMode="adjustResize">
 
             <intent-filter>
