+++ app/src/main/java/com/example/pyculator/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.pyculator
 
 import android.content.Context
 import android.os.Bundle
@@ -16,18 +16,18 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
 import com.chaquo.python.Python
 import com.chaquo.python.android.PyApplication
-import com.example.myapplication.ui.theme.PyculatorTheme
-import com.example.myapplication.pages.Page0
-import com.example.myapplication.pages.Page1
-import com.example.myapplication.pages.Page2
-import com.example.myapplication.pages.Page3
-import com.example.myapplication.viewmodels.FavoriteElement
+import com.example.pyculator.ui.theme.PyculatorTheme
+import com.example.pyculator.pages.Page0
+import com.example.pyculator.pages.Page1
+import com.example.pyculator.pages.Page2
+import com.example.pyculator.pages.Page3
+import com.example.pyculator.viewmodels.FavoriteElement
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.HorizontalPager
 import com.google.accompanist.pager.rememberPagerState
 import java.io.File
-import com.example.myapplication.viewmodels.MemoryViewModel
-import com.example.myapplication.viewmodels.SettingsViewModel
+import com.example.pyculator.viewmodels.MemoryViewModel
+import com.example.pyculator.viewmodels.SettingsViewModel
 import kotlinx.coroutines.launch
