+++ app/src/main/java/com/example/pyculator/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.theme
+package com.example.pyculator.ui.theme
 
 import androidx.compose.ui.graphics.Color
 /*
