+++ app/src/main/java/com/example/pyculator/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.theme
+package com.example.pyculator.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
